SRF구현... 근데 왜 대기시간이 더 길지..?

+++ process.h
@@ -56,8 +56,11 @@ void taskrun( Process * proc)
 //process 처리함, 소모한 시간을 돌려줌
 TIMETYPE taskupdate( Process * proc, TIMETYPE delta )
 {
-	kill(proc->pid, SIGCONT);
-	proc->state = TASK_RUN;
+	if( proc->state != TASK_RUN )
+	{
+		kill(proc->pid, SIGCONT);
+		proc->state = TASK_RUN;
+	}
 	proc->remaintime -= delta;
 	
 	//태스크가 작업을 완전히 완료 함
